+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Exported functions now live on `Wasmtime::Instance#exports` instead of `#funcs`.
 - Exports hash is now keyed by strings instead of symbols.
+- Added support for `Wasmtime::Memory` exports with `#data_size`, `#size` and `#grow` methods.
 
 ## [0.1.0] - 2020-05-07